check if tmpfile exists before delete it

+++ core/src/main/java/org/elasticsearch/index/translog/Translog.java
@@ -231,7 +231,10 @@ public class Translog extends AbstractIndexShardComponent implements IndexShardC
             }
             if (tempFileRenamed == false) {
                 try {
-                    Files.delete(tempFile);
+                    if (Files.exists(tempFile)) {
+                        Files.delete(tempFile);
+                    }
+                    
                 } catch (IOException ex) {
                     logger.warn("failed to delete temp file {}", ex, tempFile);
                 }
